Bump project versions; downgrade Im* package dependencies
Incremented project versions to 1.1.15 for ImGuizmo, ImNodes, and ImPlot. Downgraded Hexa.NET.ImGuizmo, Hexa.NET.ImNodes, and Hexa.NET.ImPlot dependencies from 2.2.8.5 to 2.2.8.4.

Add ExampleImNodes sample and improve core APIs

- Add ExampleImNodes project: idle game using KittyUI node editor
- Enhance navigation: support args in INavigation/IPage, track history with args, add ClearHistory
- Refactor node editor: better init, rename GetOuput→GetOutput, remove unused traversal code
- Update SDL/windowing: use handle APIs, improve Linux/Wayland/X11 support, fix mouse in title bar
- Update Hexa.NET package versions for compatibility
- Minor bug fixes, code style, and comments

+++ Hexa.NET.KittyUI/Application.cs
@@ -97,7 +97,7 @@
             SDL.SetHint(SDL.SDL_HINT_MOUSE_NORMAL_SPEED_SCALE, "1");
             SDL.SetHint(SDL.SDL_HINT_MOUSE_AUTO_CAPTURE, "0");
 
-            SDL.Init(InitFlags);
+            SDL.Init((uint)InitFlags);
 
             SdlCheckError();
